refactor(core): the display driver functions' API has been reviewed. The data types have been adapted to fit the actual data passed/used/returned. Since the backlight driver has undergone the similar change, the display driver code has been adapted to respect the changes.
[no changelog]

+++ core/embed/io/display/ltdc_dsi/display_internal.h
@@ -52,7 +52,7 @@ typedef struct {
   // Current display orientation (0, 90, 180, 270)
   int orientation_angle;
   // Current backlight level ranging from 0 to 255
-  int backlight_level;
+  uint8_t backlight_level;
   // The current frame buffer selector
 
   DSI_HandleTypeDef hlcd_dsi;
